Add working example with dummy wrapper values
This commit introduces a "working" example of
McMini managing state changes during execution.
Here, we simply insert a value into the mailbox
and assume that the thread under the control of libmcmini.so
properly directed the thread to write into the correct spot.
Subsequent commits will add the coordination between
libmcmini.so and McMini.