Add shared memory synchronization between libmcmini + McMini
This commit introduces synchronization between libmcmini and
the McMini process using the pair of binary semaphores in
shared memory. This is equivalent to the current model.

However, there is one subtlety. If execvp fails, we currently
block indefinitely as the the error is not propagated to the
the parent McMini process. We can resolve this using a pipe(2)
with FD_CLOEXEC. This will be introduced later