Add pipe(2) to process execvp errors in fork_process_source
This commit introduces a pipe(2) call used to
communicate errno codes from execvp in the forked child
process to the parent process. The parent uses the
presence of a value as an indication of error. When
the child successfully runs `execvp()`, and since the
write end of the pipe has been marked `FD_CLOEXEC`, the
parent will return from `read(2)` with a value of 0
and will continue execution as normal