Replace `exit()` with `quick_exit()` after forking
C++ destroys global static objects upon the program exiting
normally, i.e. with `exit(2)`. `fork_process_source` produces
new processes by `fork(2)`-ing the McMini process and attempting
to `execvp(2)` into some target process to create a template. This
is all fine when the execvp() succeeds; however, on failure, we need
be careful. Since we forked the C++ McMini process, if we simply call
`exit(2)` in the child process, this will invoke the destructors for
the C++ global static objects _which reside in the address space of
the child_. This is unexpected though: the child in-between should not
perform any clean up (that's the job of the parent)  and is only designed to
immediately exec() or else report an error via its pipe.

The solution is to use `quick_exit()` in this temporary child which does not perform the
normal clean up