Ignore SIG_IGN to SIGCHLD in libmcmini
This commit adds a `sigaction(2)` call to
ignore the SIGCHLD signal sent to the template
process. This means that the zombies of the children
will not be left around. Handling segfaults and other
issues will be a little more difficult though,
as the children of the template are NOT children
of the McMini process that contains the model